backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
psycopg[binary]>=3.1
pydantic>=2.6
pydantic-settings>=2.1
passlib[bcrypt]>=1.7
email-validator>=2.1
python-jose[cryptography]>=3.3
python-dotenv>=1.0
alembic>=1.13